+++ src/debug_utils.jl
@@ -338,6 +338,11 @@ function conditioned_varnames(context)
 end
 
 function check_varnames_seen(varnames_seen::AbstractDict{VarName,Int})
+    if isempty(varnames_seen)
+        @warn "The model does not contain any parameters."
+        return true
+    end
+
     issuccess = true
     for (varname, count) in varnames_seen
         if count == 0